+++ rxandroidble/src/main/java/com/polidea/rxandroidble/internal/RxBleLog.java
@@ -1,17 +1,11 @@
 package com.polidea.rxandroidble.internal;
 
 import android.support.annotation.IntDef;
+import android.support.annotation.Nullable;
 import android.util.Log;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.OutputStreamWriter;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.Locale;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -35,20 +29,63 @@ public class RxBleLog {
     private static final Pattern ANONYMOUS_CLASS = Pattern.compile("\\$\\d+$");
     private static final ThreadLocal<String> NEXT_TAG = new ThreadLocal<>();
 
+    private static Logger logcatLogger = new Logger() {
+        @Override
+        public void log(final int level, final String tag, final String msg) {
+            Log.println(level, tag, msg);
+        }
+    };
+
     private static int logLevel = Integer.MAX_VALUE;
-    private static File logDestination = null;
+
+    private static Logger logger = logcatLogger;
 
     private RxBleLog() {
 
     }
 
-    public static void setLogLevel(@LogLevel int logLevel) {
-        RxBleLog.logLevel = logLevel;
+    /**
+     * Simple logging interface for log messages from RxAndroidBle
+     *
+     * @see #setLogger(Logger)
+     */
+    public interface Logger {
+
+        /**
+         * @param level one of {@link Log#VERBOSE}, {@link Log#DEBUG},{@link Log#INFO},
+         *              {@link Log#WARN},{@link Log#ERROR}
+         * @param tag   log tag, caller
+         * @param msg   message to log
+         */
+        void log(int level, String tag, String msg);
     }
 
-    public static void setLogLevelWithFile(@LogLevel int logLevel, File logDestination) {
+    /**
+     * Set a custom logger implementation, set it to {@code null} to use default logcat logging
+     *
+     * Example how to forward logs to Timber:<br>
+     *
+     * <code>
+     * <pre>
+     * RxBleLog.setLogger(new RxBleLog.Logger() {
+     *    &#64;Override
+     *    public void log(final int level, final String tag, final String msg) {
+     *        Timber.tag(tag).log(level, msg);
+     *    }
+     * });
+     * </pre>
+     * </code>
+     */
+    public static void setLogger(@Nullable final Logger logger) {
+        if (logger == null) {
+            RxBleLog.logger = logcatLogger;
+        } else {
+            RxBleLog.logger = logger;
+        }
+    }
+
+    public static void setLogLevel(@LogLevel int logLevel) {
         RxBleLog.logLevel = logLevel;
-        RxBleLog.logDestination = logDestination;
     }
 
     private static String createTag() {
@@ -119,6 +156,10 @@ public class RxBleLog {
     }
 
     private static void throwShade(int priority, Throwable t, String message, Object... args) {
+        if (priority < logLevel) {
+            return;
+        }
+
         final String formattedMessage = formatString(message, args);
         final String finalMessage;
 
@@ -141,48 +182,14 @@ public class RxBleLog {
 
     private static void println(int priority, String tag, String message) {
         if (message.length() < 4000) {
-            if (!(priority < logLevel)) {
-                Log.println(priority, tag, message);
-            }
-
-            if (logDestination != null) {
-                writeToFile(logDestination, String.format(Locale.getDefault(), "%s %s", tag, message));
-            }
+            logger.log(priority, tag, message);
         } else {
             // It's rare that the message will be this large, so we're ok with the perf hit of splitting
             // and calling Log.println N times.  It's possible but unlikely that a single line will be
             // longer than 4000 characters: we're explicitly ignoring this case here.
             String[] lines = message.split("\n");
             for (String line : lines) {
-                if (!(priority < logLevel)) {
-                    Log.println(priority, tag, line);
-                }
-
-                if (logDestination != null) {
-                    writeToFile(logDestination, String.format(Locale.getDefault(), "%s %s", tag, line));
-                }
-            }
-        }
-    }
-
-    private static void writeToFile(File file, String msg) {
-        OutputStreamWriter outputStreamWriter = null;
-        try {
-            outputStreamWriter = new OutputStreamWriter(new FileOutputStream(file, true));
-            long currentTime = System.currentTimeMillis();
-            SimpleDateFormat format = new SimpleDateFormat("HH:mm:ss dd/MM/yyyy", Locale.getDefault());
-            String message = String.format(Locale.getDefault(), "%s %s.", format.format(new Date(currentTime)), msg);
-            outputStreamWriter.write(message);
-            outputStreamWriter.write("\n");
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
-            if (outputStreamWriter != null) {
-                try {
-                    outputStreamWriter.close();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
+                logger.log(priority, tag, line);
             }
         }
     }