+++ rxandroidble/src/main/java/com/polidea/rxandroidble/internal/connection/ServiceDiscoveryManager.java
@@ -6,10 +6,10 @@ import android.bluetooth.BluetoothGattService;
 import android.support.annotation.NonNull;
 
 import com.polidea.rxandroidble.RxBleDeviceServices;
-import com.polidea.rxandroidble.internal.RxBleRadio;
 import com.polidea.rxandroidble.internal.operations.OperationsProvider;
-import com.polidea.rxandroidble.internal.operations.RxBleRadioOperationServicesDiscover;
+import com.polidea.rxandroidble.internal.operations.ServiceDiscoveryOperation;
 import com.polidea.rxandroidble.internal.operations.TimeoutConfiguration;
+import com.polidea.rxandroidble.internal.serialization.ConnectionOperationQueue;
 
 import java.util.List;
 import java.util.concurrent.TimeUnit;
@@ -28,7 +28,7 @@ import rx.subjects.SerializedSubject;
 @ConnectionScope
 class ServiceDiscoveryManager {
 
-    private final RxBleRadio rxBleRadio;
+    private final ConnectionOperationQueue operationQueue;
     private final BluetoothGatt bluetoothGatt;
     private final OperationsProvider operationProvider;
     private Observable<RxBleDeviceServices> deviceServicesObservable;
@@ -42,8 +42,8 @@ class ServiceDiscoveryManager {
 
 
     @Inject
-    ServiceDiscoveryManager(RxBleRadio rxBleRadio, BluetoothGatt bluetoothGatt, OperationsProvider operationProvider) {
-        this.rxBleRadio = rxBleRadio;
+    ServiceDiscoveryManager(ConnectionOperationQueue operationQueue, BluetoothGatt bluetoothGatt, OperationsProvider operationProvider) {
+        this.operationQueue = operationQueue;
         this.bluetoothGatt = bluetoothGatt;
         this.operationProvider = operationProvider;
         reset();
@@ -109,17 +109,17 @@ class ServiceDiscoveryManager {
         return new Func1<TimeoutConfiguration, Observable<RxBleDeviceServices>>() {
             @Override
             public Observable<RxBleDeviceServices> call(TimeoutConfiguration timeoutConf) {
-                final RxBleRadioOperationServicesDiscover operation = operationProvider
+                final ServiceDiscoveryOperation operation = operationProvider
                         .provideServiceDiscoveryOperation(timeoutConf.timeout, timeoutConf.timeoutTimeUnit);
                 if (addDelay) {
                     return Observable.timer(delay, timeUnit).flatMap(new Func1<Long, Observable<RxBleDeviceServices>>() {
                         @Override
                         public Observable<RxBleDeviceServices> call(Long aLong) {
-                            return rxBleRadio.queue(operation);
+                            return operationQueue.queue(operation);
                         }
                     });
                 } else {
-                    return rxBleRadio.queue(operation);
+                    return operationQueue.queue(operation);
                 }
             }
         };